+++ sql/emaj_prepare_parallel_rollback_test.sql
@@ -11,7 +11,7 @@
 --      VALUES ('dblink_user_password','user=<user> password=<password>'
 --
 -- Once this script is executed, you can type:
--- ./client/emajParallelRollback.php -g 'emaj parallel rollback test group' -m BEFORE_PROC_2 -s 3 -v <and connection parameters>
+-- ./client/emajParallelRollback.pl -g 'emaj parallel rollback test group' -m BEFORE_PROC_2 -s 3 -v <and connection parameters>
 --
 -- The role configured for the dblink connection and the role used to perform the parallel 
 -- rollback operation must also be at least member of emaj_adm.
@@ -247,9 +247,9 @@ RESET search_path;
 
 \echo '---'
 \echo '--- emaj_prepare_parallel_rollback_test.sql script successfuly completed.'
-\echo '--- A parallel rollback can be performed, from the E-Maj main directory, using a command like:'
-\echo '    ./client/emajParallelRollback.php -g ''emaj parallel rollback test group'' -m BEFORE_PROC_2 -s 3 -v <and any needed connection parameters among -dhpUW>'
-\echo '--- "./client/emajParallelRollback.php --help" provides information about available parameters.'
+\echo '--- A parallel rollback can be performed, using a command like:'
+\echo '--- emajParallelRollback.pl -g ''emaj parallel rollback test group'' -m BEFORE_PROC_2 -s 3 -v <and any needed connection parameters among -dhpUW>'
+\echo '--- emajParallelRollback.pl --help" provides information about available parameters.'
 \echo '---'
 
 \unset ECHO